backend/requirements.txt
Flask>=2.2
SQLAlchemy>=2.0
Flask-SQLAlchemy>=3.0
Flask-Restless-NG>=3.0.0
PyMySQL>=1.1.0
cryptography>=41.0.0
Flask-Cors>=4.0.0
python-dotenv
pytest